refactor: consolidate handler functions into separate files for better organization
