+++ cmd/web/handlers.go
@@ -1,0 +1,32 @@
+package main
+
+import (
+	"fmt"
+	"net/http"
+	"strconv"
+)
+
+func home(w http.ResponseWriter, r *http.Request) {
+	w.Header().Add("Server", "Go")
+	w.Write([]byte("Hello from snippetbox"))
+}
+
+func snippetView(w http.ResponseWriter, r *http.Request) {
+	id, err := strconv.Atoi(r.PathValue("id"))
+	if err != nil || id < 1 {
+		http.NotFound(w, r)
+		return
+	}
+	fmt.Fprintf(w, "Displaying a sp ecific snippet with ID %d...", id)
+}
+
+func snippetCreate(w http.ResponseWriter, r *http.Request) {
+	w.Write([]byte("Create a snippet"))
+}
+
+func snippetCreatePost(w http.ResponseWriter, r *http.Request) {
+	// send a 201 status code.
+	w.WriteHeader(http.StatusCreated)
+
+	w.Write([]byte("save a new snippet"))
+}
